# server/swagger.yaml
openapi: 3.0.0
info:
  title: BookEase API
  version: 1.0.0
  description: REST API for BookEase Appointment Booking System

servers:
  - url: http://localhost:3000/api
    description: Local Development Server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        location:
          type: string
        role:
          type: string
          enum: [admin, user]
        created_at:
          type: string

    Service:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        duration_minutes:
          type: integer
        price:
          type: number
        image_url:
          type: string
        created_at:
          type: string

    Appointment:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        service_id:
          type: integer
        appointment_date:
          type: string
        appointment_time:
          type: string
        status:
          type: string
          enum: [pending, confirmed, cancelled, completed]
        notes:
          type: string
        created_at:
          type: string

paths:

  # ==================== AUTH ====================
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, phone, location, password]
              properties:
                name:
                  type: string
                  example: Juan Dela Cruz
                email:
                  type: string
                  example: juan@email.com
                phone:
                  type: string
                  example: 09123456789
                location:
                  type: string
                  example: Cebu City
                password:
                  type: string
                  example: password123
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Registered successfully
        '400':
          description: Email already exists or missing fields
        '500':
          description: Server error

  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login user and get JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  example: juan@email.com
                password:
                  type: string
                  example: password123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid password
        '404':
          description: User not found
        '500':
          description: Server error

  # ==================== SERVICES ====================
  /services:
    get:
      tags:
        - Services
      summary: Get all services (Public)
      responses:
        '200':
          description: List of all services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Service'
        '500':
          description: Server error

    post:
      tags:
        - Services
      summary: Create a new service (Admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [name, description, duration_minutes, price]
              properties:
                name:
                  type: string
                  example: General Consultation
                description:
                  type: string
                  example: Basic doctor consultation
                duration_minutes:
                  type: integer
                  example: 30
                price:
                  type: number
                  example: 500.00
                image:
                  type: string
                  format: binary
      responses:
        '201':
          description: Service created successfully
        '403':
          description: Admin access required
        '500':
          description: Server error

  /services/{id}:
    put:
      tags:
        - Services
      summary: Update a service (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                duration_minutes:
                  type: integer
                price:
                  type: number
      responses:
        '200':
          description: Service updated successfully
        '403':
          description: Admin access required
        '500':
          description: Server error

    delete:
      tags:
        - Services
      summary: Delete a service (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Service deleted successfully
        '403':
          description: Admin access required
        '500':
          description: Server error

  # ==================== APPOINTMENTS ====================
  /appointments:
    get:
      tags:
        - Appointments
      summary: Get all appointments (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, confirmed, cancelled, completed]
          description: Filter by status
        - in: query
          name: search
          schema:
            type: string
          description: Search by patient or service name
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of all appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Appointment'
        '403':
          description: Admin access required
        '500':
          description: Server error

    post:
      tags:
        - Appointments
      summary: Book a new appointment (User)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [service_id, appointment_date, appointment_time]
              properties:
                service_id:
                  type: integer
                  example: 1
                appointment_date:
                  type: string
                  example: "2026-05-10"
                appointment_time:
                  type: string
                  example: "09:00"
                notes:
                  type: string
                  example: Please prepare medical records
      responses:
        '201':
          description: Appointment booked successfully
        '401':
          description: No token provided
        '500':
          description: Server error

  /appointments/my:
    get:
      tags:
        - Appointments
      summary: Get current user appointments
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, confirmed, cancelled, completed]
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: User appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Appointment'
        '401':
          description: Unauthorized
        '500':
          description: Server error

  /appointments/check:
    get:
      tags:
        - Appointments
      summary: Check booked time slots for a date and service
      parameters:
        - in: query
          name: date
          required: true
          schema:
            type: string
          example: "2026-05-10"
        - in: query
          name: service_id
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: List of booked time slots
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    appointment_time:
                      type: string
        '400':
          description: Missing required parameters
        '500':
          description: Server error

  /appointments/{id}/status:
    patch:
      tags:
        - Appointments
      summary: Update appointment status (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [pending, confirmed, completed, cancelled]
                  example: confirmed
      responses:
        '200':
          description: Status updated successfully
        '403':
          description: Admin access required
        '500':
          description: Server error

  /appointments/{id}:
    delete:
      tags:
        - Appointments
      summary: Delete an appointment
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Appointment deleted successfully
        '401':
          description: Unauthorized
        '500':
          description: Server error

  # ==================== USERS ====================
  /users/profile:
    get:
      tags:
        - Users
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
        '500':
          description: Server error

    put:
      tags:
        - Users
      summary: Update current user profile (with optional image upload)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Juan Dela Cruz
                phone:
                  type: string
                  example: 09123456789
                location:
                  type: string
                  example: Cebu City
                profile_image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Profile updated successfully
        '401':
          description: Unauthorized
        '500':
          description: Server error

  /users:
    get:
      tags:
        - Users
      summary: Get all users (Admin only)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of all users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '403':
          description: Admin access required
        '500':
          description: Server error